Conflicts:
	src/main/webapp/departement/departement.xhtml

+++ src/main/java/com/douwe/notes/web/beans/AnneeAcademiqueBean.java
@@ -4,13 +4,14 @@ import com.douwe.notes.entities.AnneeAcademique;
 import com.douwe.notes.service.IAnneeAcademiqueService;
 import static java.awt.SystemColor.text;
 import java.util.Iterator;
+import java.util.LinkedList;
 import java.util.List;
 import javax.ejb.EJB;
 import javax.inject.Named;
-import javax.enterprise.context.RequestScoped; 
-import javax.faces.application.FacesMessage;  
-import javax.faces.context.FacesContext;  
-import javax.faces.event.ActionEvent; 
+import javax.enterprise.context.RequestScoped;
+import javax.faces.application.FacesMessage;
+import javax.faces.context.FacesContext;
+import javax.faces.event.ActionEvent;
 
 @Named(value = "anneeAcademiqueBean")
 @RequestScoped
@@ -19,52 +20,51 @@ public class AnneeAcademiqueBean {
     @EJB
     private IAnneeAcademiqueService service;
     private AnneeAcademique anneeAcademique = new AnneeAcademique();
-    private List<AnneeAcademique> anneeAcademiques;     
+    private List<AnneeAcademique> anneeAcademiques;
     private String message;
 
     /**
      * Creates a new instance of AnneeAcademiqueBean
      */
-    public AnneeAcademiqueBean() {        
-        message="";
+    public AnneeAcademiqueBean() {
+        message = "";
     }
 
-
     public String saveOrUpdateAnneeAcademique() {
-        if (anneeAcademique != null) {            
+        if (anneeAcademique != null&& anneeAcademique.getDebut().getYear()<= anneeAcademique.getFin().getYear()&& anneeAcademique.getDebut().getMonth()<=anneeAcademique.getFin().getMonth() && anneeAcademique.getDebut().getDay()<anneeAcademique.getFin().getDay()) {
             service.saveOrUpdateAnnee(anneeAcademique);
-            anneeAcademique = new AnneeAcademique();                      
-        }
+            anneeAcademique = new AnneeAcademique();
+        }        
         return "saveOrUpdateAnneeAcademique";
     }
 
     public String deleteAnneeAcademique() {
-        if (anneeAcademique != null) {          
+        if (anneeAcademique != null&& anneeAcademique.getDebut().getYear()<= anneeAcademique.getFin().getYear()&& anneeAcademique.getDebut().getMonth()<=anneeAcademique.getFin().getMonth() && anneeAcademique.getDebut().getDay()<anneeAcademique.getFin().getDay()) {
             message = "Suppression reussi";
             service.deleteAnnee(anneeAcademique.getId());
-            anneeAcademique = new AnneeAcademique();            
-        }
+            anneeAcademique = new AnneeAcademique();
+        }        
         return "deleteAnneeAcademique";
     }
 
     public String choix(int n) {
         if (n == 1) {
-            anneeAcademique=new AnneeAcademique();
-            message="Enregistrement reussi ";
+            anneeAcademique = new AnneeAcademique();
+            message = "Enregistrement reussi ";
             return "saveAnneeAcademique";
-        }
-
-        else if (n == 2 && anneeAcademique!=null &&anneeAcademique.getVersion() >= 1) {
-            message="Mise à jour reussi ";
+        } else if (n == 2 && anneeAcademique != null && anneeAcademique.getVersion() >= 1) {
+            message = "Mise à jour reussi ";
             return "updateAnneeAcademique";
         }
-        anneeAcademique = new AnneeAcademique();        
+        anneeAcademique = new AnneeAcademique();
         return "anneeAcademique";
     }
-    public void notification(ActionEvent actionEvent) {  
-        FacesContext context = FacesContext.getCurrentInstance();            
-        context.addMessage(null, new FacesMessage("Succes", message));          
+
+    public void notification(ActionEvent actionEvent) {
+        FacesContext context = FacesContext.getCurrentInstance();
+        context.addMessage(null, new FacesMessage("Succes", message));
     }
+
     public IAnneeAcademiqueService getService() {
         return service;
     }
@@ -73,21 +73,24 @@ public class AnneeAcademiqueBean {
         this.service = service;
     }
 
-
     public AnneeAcademique getAnneeAcademique() {
         return anneeAcademique;
     }
 
-    public void setAnneeAcademique(AnneeAcademique anneeAcademique) {        
+    public void setAnneeAcademique(AnneeAcademique anneeAcademique) {
         this.anneeAcademique = anneeAcademique;
     }
 
-    public List<AnneeAcademique> getAnneeAcademiques() {        
-        anneeAcademiques = service.getAllAnnee();        
+    public List<AnneeAcademique> getAnneeAcademiques() {
+        anneeAcademiques = service.getAllAnnee();
         return anneeAcademiques;
     }
 
     public void setAnneeAcademiques(List<AnneeAcademique> anneeAcademiques) {
         this.anneeAcademiques = anneeAcademiques;
+    }
+
+    public int inc(int n) {
+        return ++n;
     }
 }
